BuddyAI/backend/requirements.txt
# Django and core dependencies
Django>=5.2.4
python-dotenv>=1.0.0

# LangChain and LLM dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.0
openai>=1.97.0

# Vector store and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Text processing
nltk>=3.8
tiktoken>=0.5.0

# PDF processing
PyPDF2>=3.0.0

# Additional utilities
rank-bm25>=0.2.2 